pyproject.toml
[project]
name = "mini-leaderboard"
version = "0.0.1"
description = "simple learder board with many projects"
authors = [{ name = "Wh1isper", email = "jizhongsheng957@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "alembic>=1.15.2",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg[binary,pool]>=3.2.8",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]

[project.urls]
Homepage = "https://wh1isper.github.io/mini-leaderboard/"
Repository = "https://github.com/wh1isper/mini-leaderboard"
Documentation = "https://wh1isper.github.io/mini-leaderboard/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "pytest-asyncio>=0.25.3",
    "tox-uv>=1.11.3",
    "deptry>=0.22.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.9.2",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "docker",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
py-modules = ["mini_leaderboard"]

[tool.mypy]
files = ["mini_leaderboard"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    "TRY300",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "E722"]
"mini_leaderboard/dbutils.py" = ["S603", "S607", "B008"]
"mini_leaderboard/alembic/env.py" = ["A002", "SIM", "E"]
"mini_leaderboard/config.py" = ["E402"]

[tool.deptry.per_rule_ignores]
DEP002 = ["httpx", "psycopg"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["mini_leaderboard"]

[project.scripts]
mini-leaderboard = "mini_leaderboard.cli:cli"
